assign/Dravin_new_assign/chatbot-suite/backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic_settings
motor